ft_atoi restored, I created a folder utils with the custom version

+++ lib/libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: ede-smet <ede-smet@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/09 17:00:04 by ede-smet          #+#    #+#             */
-/*   Updated: 2023/02/25 14:53:23 by ede-smet         ###   ########.fr       */
+/*   Updated: 2023/02/25 15:08:16 by ede-smet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long int	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-	int				i;
-	int				sign;
-	long long int	to_i;
+	int	i;
+	int	sign;
+	int	to_i;
 
 	to_i = 0;
 	sign = 1;
@@ -29,7 +29,7 @@ long long int	ft_atoi(const char *str)
 			sign *= -1;
 		i++;
 	}
-	while (ft_isdigit(str[i]))
+	while (ft_isdigit((int)str[i]))
 	{
 		to_i = to_i * 10 + (str[i] - '0');
 		i++;
